# go-core/api/openapi.yaml
openapi: 3.0.3
info:
  title: Authorization Engine REST API
  description: |
    Comprehensive REST API for the Authorization Engine providing policy-based access control,
    principal management, and policy import/export functionality.

    ## Features
    - **Authorization Checks**: Validate permissions for principals and resources
    - **Policy Management**: Full CRUD operations for authorization policies
    - **Principal Management**: Manage users, roles, and attributes
    - **Export/Import**: Backup, restore, and migrate policies
    - **Versioning**: API versioning with backward compatibility

    ## Authentication
    All endpoints (except `/health`) require JWT Bearer token authentication.
    Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```

    ## Rate Limiting
    - 100 requests per minute per IP address
    - 1000 requests per hour per authenticated user
    - Rate limit headers included in all responses

    ## Versioning
    Current version: v1
    Base path: `/v1`

  version: 1.0.0
  contact:
    name: API Support
    email: support@authz-engine.example.com
    url: https://github.com/authz-engine/go-core
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: https://api.authz-engine.example.com/v1
    description: Production server
  - url: https://staging-api.authz-engine.example.com/v1
    description: Staging server
...
security:
  - BearerAuth: []
paths:
...
  /authorization/check:
    post:
      tags:
        - Authorization
      summary: Check authorization
      description: |
        Check if a principal is authorized to perform an action on a resource.
        Returns a boolean decision with optional debug information.
      operationId: checkAuthorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthorizationCheckRequest'
            examples:
              simple:
                summary: Simple authorization check
                value:
                  principal:
                    id: "user123"
                    roles: ["developer"]
                  action: "read"
                  resource:
                    kind: "document"
                    id: "doc456"
              with_attributes:
                summary: Check with principal attributes
                value:
                  principal:
                    id: "user123"
                    roles: ["developer", "team-lead"]
                    attributes:
                      department: "engineering"
                      level: "senior"
                      location: "US"
                  action: "write"
                  resource:
                    kind: "repository"
                    id: "repo789"
                    attributes:
                      visibility: "private"
                      team: "backend"
              with_context:
                summary: Check with request context
                value:
                  principal:
                    id: "user123"
                    roles: ["developer"]
                  action: "deploy"
                  resource:
                    kind: "application"
                    id: "app001"
                  context:
                    ip_address: "192.168.1.100"
                    time: "2025-01-27T10:30:00Z"
                    environment: "production"
      responses:
        '200':
          description: Authorization decision
          headers:
            X-RateLimit-Limit:
              schema:
                type: integer
              description: Request limit per window
            X-RateLimit-Remaining:
              schema:
                type: integer
              description: Remaining requests in current window
            X-RateLimit-Reset:
              schema:
                type: integer
              description: Unix timestamp when the rate limit resets
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthorizationCheckResponse'
              examples:
                allowed:
                  summary: Access granted
                  value:
                    allowed: true
                    decision_time_ms: 5
                    matched_policies:
                      - "policy-dev-read-docs"
                denied:
                  summary: Access denied
                  value:
                    allowed: false
                    decision_time_ms: 3
                    reason: "No matching policy found"
                denied_explicit:
                  summary: Explicitly denied
                  value:
                    allowed: false
                    decision_time_ms: 4
                    reason: "Denied by policy: prod-deny-deploy"
                    matched_policies:
                      - "policy-prod-deny-deploy"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /authorization/check-resources:
    post:
      tags:
        - Authorization
      summary: Batch check multiple resources
      description: |
        Check authorization for a principal across multiple resources.
        Returns a list of decisions, one for each resource.
      operationId: checkResourcesAuthorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchAuthorizationCheckRequest'
            example:
              principal:
                id: "user123"
                roles: ["developer"]
              action: "read"
              resources:
                - kind: "document"
                  id: "doc456"
                - kind: "document"
                  id: "doc789"
                - kind: "repository"
                  id: "repo123"
      responses:
        '200':
          description: Batch authorization decisions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchAuthorizationCheckResponse'
              example:
                results:
                  - resource:
                      kind: "document"
                      id: "doc456"
                    allowed: true
                    decision_time_ms: 3
                  - resource:
                      kind: "document"
                      id: "doc789"
                    allowed: false
                    decision_time_ms: 2
                    reason: "No matching policy"
                  - resource:
                      kind: "repository"
                      id: "repo123"
                    allowed: true
                    decision_time_ms: 4
                total_time_ms: 9
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /authorization/allowed-actions:
    get:
      tags:
        - Authorization
      summary: Get allowed actions
      description: |
        Get all actions that a principal is allowed to perform on a resource.
        Useful for UI rendering and permission discovery.
      operationId: getAllowedActions
      parameters:
        - name: principal_id
          in: query
          required: true
          schema:
            type: string
          description: Principal identifier
          example: "user123"
        - name: resource_kind
          in: query
          required: true
          schema:
            type: string
          description: Resource kind
          example: "document"
        - name: resource_id
          in: query
          required: true
          schema:
            type: string
          description: Resource identifier
          example: "doc456"
        - name: roles
          in: query
          required: false
          schema:
            type: array
            items:
              type: string
          style: form
          explode: true
          description: Principal roles
          example: ["developer", "admin"]
      responses:
        '200':
          description: List of allowed actions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AllowedActionsResponse'
              example:
                allowed_actions:
                  - "read"
                  - "write"
                  - "comment"
                denied_actions:
                  - "delete"
                  - "share"
                decision_time_ms: 8
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /policies:
    get:
      tags:
        - Policies
      summary: List policies
      description: |
        List all policies with optional filtering and pagination.
        Supports filtering by resource kind, principal roles, and policy type.
      operationId: listPolicies
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number
        - name: page_size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Number of items per page
        - name: resource_kind
          in: query
          schema:
            type: string
          description: Filter by resource kind
          example: "document"
        - name: principal_role
          in: query
          schema:
            type: string
          description: Filter by principal role
          example: "developer"
        - name: policy_type
          in: query
          schema:
            type: string
            enum: [resource, principal, derived_role]
          description: Filter by policy type
        - name: sort
          in: query
          schema:
            type: string
            enum: [id, created_at, updated_at]
            default: created_at
          description: Sort field
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
          description: Sort order
      responses:
        '200':
          description: List of policies
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyListResponse'
              example:
                policies:
                  - id: "policy-dev-read-docs"
                    version: "1.0"
                    description: "Developers can read documentation"
                    resource_policy:
                      resource: "document"
                      version: "1.0"
                      rules:
                        - actions: ["read"]
                          effect: "EFFECT_ALLOW"
                          roles: ["developer"]
                    metadata:
                      created_at: "2025-01-20T10:00:00Z"
                      updated_at: "2025-01-20T10:00:00Z"
                      created_by: "admin"
                  - id: "policy-admin-all"
                    version: "1.0"
                    description: "Admins have full access"
                    principal_policy:
                      principal: "role:admin"
                      version: "1.0"
                      rules:
                        - resource: "*"
                          actions:
                            - name: "*"
                              effect: "EFFECT_ALLOW"
                    metadata:
                      created_at: "2025-01-20T09:00:00Z"
                      updated_at: "2025-01-20T09:00:00Z"
                      created_by: "admin"
                pagination:
                  page: 1
                  page_size: 20
                  total_items: 150
                  total_pages: 8
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Policies
      summary: Create policy
      description: |
        Create a new authorization policy.
        Supports resource policies, principal policies, and derived roles.
      operationId: createPolicy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Policy'
            examples:
              resource_policy:
                summary: Resource policy
                value:
                  id: "policy-dev-read-docs"
                  version: "1.0"
                  description: "Developers can read documentation"
                  resource_policy:
                    resource: "document"
                    version: "1.0"
                    rules:
                      - actions: ["read", "list"]
                        effect: "EFFECT_ALLOW"
                        roles: ["developer"]
                        condition:
                          match:
                            expr: "resource.attr.visibility == 'public'"
              principal_policy:
                summary: Principal policy
                value:
                  id: "policy-user123-admin"
                  version: "1.0"
                  description: "User123 admin permissions"
                  principal_policy:
                    principal: "user123"
                    version: "1.0"
                    rules:
                      - resource: "document"
                        actions:
                          - name: "*"
                            effect: "EFFECT_ALLOW"
              derived_role:
                summary: Derived role
                value:
                  id: "derived-role-senior-dev"
                  version: "1.0"
                  description: "Senior developer derived role"
                  derived_roles:
                    name: "senior_developer"
                    definitions:
                      - name: "senior_developer"
                        parent_roles:
                          - "developer"
                        condition:
                          match:
                            expr: "P.attr.level == 'senior' && P.attr.tenure_years >= 2"
      responses:
        '201':
          description: Policy created successfully
          headers:
            Location:
              schema:
                type: string
              description: URL of the created policy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyResponse'
              example:
                id: "policy-dev-read-docs"
                version: "1.0"
                created_at: "2025-01-27T10:30:00Z"
                message: "Policy created successfully"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Policy already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: "POLICY_ALREADY_EXISTS"
                  message: "A policy with ID 'policy-dev-read-docs' already exists"
                  details:
                    existing_policy_id: "policy-dev-read-docs"
        '500':
          $ref: '#/components/responses/InternalServerError'
  /policies/{id}:
    get:
      tags:
        - Policies
      summary: Get policy by ID
      description: Retrieve a specific policy by its identifier
      operationId: getPolicy
      parameters:
        - $ref: '#/components/parameters/PolicyId'
      responses:
        '200':
          description: Policy details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Policy'
              example:
                id: "policy-dev-read-docs"
                version: "1.0"
                description: "Developers can read documentation"
                resource_policy:
                  resource: "document"
                  version: "1.0"
                  rules:
                    - actions: ["read"]
                      effect: "EFFECT_ALLOW"
                      roles: ["developer"]
                metadata:
                  created_at: "2025-01-20T10:00:00Z"
                  updated_at: "2025-01-20T10:00:00Z"
                  created_by: "admin"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags:
        - Policies
      summary: Update policy
      description: |
        Update an existing policy. The entire policy is replaced.
        To partially update, retrieve the policy first, modify it, then send the complete updated policy.
      operationId: updatePolicy
      parameters:
        - $ref: '#/components/parameters/PolicyId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Policy'
            example:
              id: "policy-dev-read-docs"
              version: "1.1"
              description: "Developers can read and comment on documentation"
              resource_policy:
                resource: "document"
                version: "1.1"
                rules:
                  - actions: ["read", "comment"]
                    effect: "EFFECT_ALLOW"
                    roles: ["developer"]
      responses:
        '200':
          description: Policy updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyResponse'
              example:
                id: "policy-dev-read-docs"
                version: "1.1"
                updated_at: "2025-01-27T10:30:00Z"
                message: "Policy updated successfully"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags:
        - Policies
      summary: Delete policy
      description: |
        Delete a policy by ID. This action is irreversible.
        Consider creating a backup before deleting important policies.
      operationId: deletePolicy
      parameters:
        - $ref: '#/components/parameters/PolicyId'
      responses:
        '204':
          description: Policy deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /policies/export:
    post:
      tags:
        - Policies
      summary: Export policies
      description: |
        Export policies in JSON or YAML format.
        Supports filtering by IDs or exporting all policies.
        Can create a complete backup bundle with metadata.
      operationId: exportPolicies
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExportPoliciesRequest'
            examples:
              export_specific:
                summary: Export specific policies
                value:
                  policy_ids:
                    - "policy-dev-read-docs"
                    - "policy-admin-all"
                  format: "json"
              export_all:
                summary: Export all policies
                value:
                  format: "yaml"
              export_bundle:
                summary: Export as bundle with metadata
                value:
                  format: "json"
                  include_metadata: true
                  include_principals: true
      responses:
        '200':
          description: Exported policies
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportPoliciesResponse'
              example:
                format: "json"
                policies:
                  - id: "policy-dev-read-docs"
                    version: "1.0"
                    description: "Developers can read documentation"
                    resource_policy:
                      resource: "document"
                      version: "1.0"
                      rules:
                        - actions: ["read"]
                          effect: "EFFECT_ALLOW"
                          roles: ["developer"]
                metadata:
                  exported_at: "2025-01-27T10:30:00Z"
                  exported_by: "admin"
                  total_policies: 1
                  format_version: "1.0"
            application/x-yaml:
              schema:
                type: string
              example: |
                format: yaml
                policies:
                  - id: policy-dev-read-docs
                    version: "1.0"
                    description: Developers can read documentation
                    resource_policy:
                      resource: document
                      version: "1.0"
                      rules:
                        - actions: [read]
                          effect: EFFECT_ALLOW
                          roles: [developer]
                metadata:
                  exported_at: "2025-01-27T10:30:00Z"
                  exported_by: admin
                  total_policies: 1
                  format_version: "1.0"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /policies/backup:
    post:
      tags:
        - Policies
      summary: Create backup
      description: |
        Create a complete backup of all policies and optionally principals.
        The backup includes metadata and can be used for restore operations.
      operationId: createBackup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBackupRequest'
            example:
              name: "daily-backup-2025-01-27"
              description: "Daily backup before maintenance"
              include_principals: true
      responses:
        '201':
          description: Backup created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BackupResponse'
              example:
                backup_id: "backup-20250127-103000"
                name: "daily-backup-2025-01-27"
                created_at: "2025-01-27T10:30:00Z"
                size_bytes: 1048576
                policy_count: 150
                principal_count: 1000
                format: "json"
                checksum: "sha256:abc123..."
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /policies/restore:
    post:
      tags:
        - Policies
      summary: Restore from backup
      description: |
        Restore policies from a backup.
        Supports dry-run mode and different restore strategies.
      operationId: restoreBackup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RestoreBackupRequest'
            examples:
              restore_full:
                summary: Full restore (replace all)
                value:
                  backup_id: "backup-20250127-103000"
                  mode: "replace_all"
              restore_merge:
                summary: Merge with existing
                value:
                  backup_id: "backup-20250127-103000"
                  mode: "merge"
                  conflict_resolution: "keep_existing"
              restore_dry_run:
                summary: Dry run validation
                value:
                  backup_id: "backup-20250127-103000"
                  mode: "replace_all"
                  dry_run: true
      responses:
        '200':
          description: Restore completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestoreBackupResponse'
              example:
                success: true
                restored_policies: 150
                restored_principals: 1000
                skipped_count: 0
                failed_count: 0
                dry_run: false
                message: "Successfully restored 150 policies and 1000 principals"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Backup not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: "BACKUP_NOT_FOUND"
                  message: "Backup with ID 'backup-20250127-103000' not found"
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT Bearer token authentication. Include the token in the Authorization header:
        ```
        Authorization: Bearer <your-jwt-token>
        ```
...
  schemas:
...
    AuthorizationCheckRequest:
      type: object
      required:
        - principal
        - action
        - resource
      properties:
        principal:
          $ref: '#/components/schemas/Principal'
        action:
          type: string
          description: Action to be performed
          example: "read"
        resource:
          $ref: '#/components/schemas/Resource'
        context:
          type: object
          additionalProperties: true
          description: Additional context for policy evaluation
          example:
            ip_address: "192.168.1.100"
            time: "2025-01-27T10:30:00Z"
            environment: "production"
    AuthorizationCheckResponse:
      type: object
      required:
        - allowed
      properties:
        allowed:
          type: boolean
          description: Whether the action is allowed
        decision_time_ms:
          type: integer
          description: Time taken to make decision in milliseconds
        matched_policies:
          type: array
          items:
            type: string
          description: List of policy IDs that matched
        reason:
          type: string
          description: Reason for denial (if not allowed)
...